Fix gui flickering when going from less than 100 to 100 caffeine level.
We simply set a minimal width of the gui so it's always at the width necessary for a
level of 100. However we need to destroy and rebuild the gui at game load because the
gui is saved between loads and would no update for savegames where the autoinjector is
already researched.